included pdfbox in lib folder and got it to run

+++ FlowReader/src/flowreader/utils/ValueComparator.java
@@ -4,6 +4,7 @@
  */
 package flowreader.utils;
 
+import java.math.BigDecimal;
 import java.util.Comparator;
 import java.util.Map;
 
@@ -13,15 +14,15 @@ import java.util.Map;
  */
 public class ValueComparator implements Comparator<String> {
 
-    Map<String, Integer> base;
-    public ValueComparator(Map<String, Integer> base) {
+    Map<String, BigDecimal> base;
+    public ValueComparator(Map<String, BigDecimal> base) {
         this.base = base;
     }
 
     // Note: this comparator imposes orderings that are inconsistent with equals.    
     @Override
     public int compare(String a, String b) {
-        if (base.get(a) >= base.get(b)) {
+        if (base.get(a).intValue() >= base.get(b).intValue()) {
             return -1;
         } else {
             return 1;
